refactor: jdbc connection 패키지 수정
