+++ src/main/java/common/JdbcConnection.java
@@ -1,4 +1,4 @@
-package common.connection;
+package common;
 
 import java.sql.Connection;
 import java.sql.DriverManager;
